+++ psi4/src/psi4/libciomr/init_matrix.cc
@@ -66,22 +66,22 @@ namespace psi {
 **
 ** \ingroup CIOMR
 */
-double ** init_matrix(unsigned long int n, unsigned long int m)
+double ** init_matrix(size_t n, size_t m)
 {
     double **A=NULL;
     double *B=NULL;
-    unsigned long int i;
+    size_t i;
 
     if(!m || !n) return(static_cast<double **>(0));
 
-//  if ((A = (double **) malloc(n * (unsigned long int)sizeof(double *)))==NULL) {
+//  if ((A = (double **) malloc(n * (size_t)sizeof(double *)))==NULL) {
     if ((A = new double*[n])==NULL) {
         outfile->Printf("block_matrix: trouble allocating memory \n");
         outfile->Printf("n = %ld\n",n);
         exit(PSI_RETURN_FAILURE);
     }
 
-//  if ((B = (double *) malloc(m*n * (unsigned long int)sizeof(double)))==NULL) {
+//  if ((B = (double *) malloc(m*n * (size_t)sizeof(double)))==NULL) {
     if ((B = new double[n*m])==NULL) {
         outfile->Printf("block_matrix: trouble allocating memory \n");
         outfile->Printf("m = %ld\n",m);
@@ -89,7 +89,7 @@ double ** init_matrix(unsigned long int n, unsigned long int m)
     }
 
     // bzero is not in the C standard, use memset instead.
-    //bzero(B, m*n*(unsigned long int)sizeof(double));
+    //bzero(B, m*n*(size_t)sizeof(double));
     memset(static_cast<void*>(B), 0, m*n*sizeof(double));
 
     for (i = 0; i < n; i++) {
@@ -103,9 +103,9 @@ double ** init_matrix(unsigned long int n, unsigned long int m)
 
   /**
   double **array=NULL;
-  unsigned long int i;
+  size_t i;
 
-  if ((array = (double **) malloc(n*(unsigned long int)sizeof(double *)))
+  if ((array = (double **) malloc(n*(size_t)sizeof(double *)))
     ==NULL) {
     outfile->Printf("init_matrix: trouble allocating memory \n");
     outfile->Printf("n = %ld\n",n);
@@ -113,13 +113,13 @@ double ** init_matrix(unsigned long int n, unsigned long int m)
   }
 
   for (i = 0; i < n; i++) {
-    if ((array[i] = (double *) malloc(m*(unsigned long int)sizeof(double)))
+    if ((array[i] = (double *) malloc(m*(size_t)sizeof(double)))
       ==NULL) {
       outfile->Printf("init_matrix: trouble allocating memory \n");
       outfile->Printf("i = %ld m = %ld\n",i,m);
       exit(PSI_RETURN_FAILURE);
     }
-    bzero(array[i],m*(unsigned long int)sizeof(double));
+    bzero(array[i],m*(size_t)sizeof(double));
   }
   return(array);
   **/
@@ -145,7 +145,7 @@ double ** init_matrix(unsigned long int n, unsigned long int m)
 **
 ** \ingroup CIOMR
 */
-void free_matrix(double **array, unsigned long int /*size*/)
+void free_matrix(double **array, size_t /*size*/)
 {
     if(array == NULL) return;
     delete [] array[0];
@@ -155,7 +155,7 @@ void free_matrix(double **array, unsigned long int /*size*/)
 // <<<<<<<<<<<<<<<<<<<<<
 
   /**
-  unsigned long int i;
+  size_t i;
 
   for (i=0; i < size ; i++) {
     free(array[i]);
